% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_and_reactions.R
\name{build_age_reactions}
\alias{build_age_reactions}
\title{get reaction list from number of age groups m and
contact_matrix}
\usage{
build_age_reactions(contact_matrix)
}
\arguments{
\item{contact_matrix}{matrix m x m}
}
\description{
get reaction list from number of age groups m and
contact_matrix
}
