% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_and_reactions.R
\name{calculate_age_structured_beta}
\alias{calculate_age_structured_beta}
\title{Using the next generation matrix approach, calculates beta based on a fixed
R0}
\usage{
calculate_age_structured_beta(sm, contact_matrix, population_sizes, R0)
}
\arguments{
\item{sm}{Object of class \code{stochastic_model}}

\item{contact_matrix}{An m x m contact matrix}

\item{population_sizes}{An m vector of population size}

\item{R0}{numeric reproduction number}
}
\value{
numeric
}
\description{
Using the next generation matrix approach, calculates beta based on a fixed
R0
}
