% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_and_reactions.R
\name{create_age_initial_conditions}
\alias{create_age_initial_conditions}
\title{Convert a vector of each state into a labeled state using
\code{age_groups} labels}
\usage{
create_age_initial_conditions(sim_scenario, age_groups)
}
\arguments{
\item{sim_scenario}{output of \code{load_model_params}}

\item{age_groups}{vector of age group labels}
}
\description{
Convert a vector of each state into a labeled state using
\code{age_groups} labels
}
\examples{
sim_scenario <- list(
  "initial_states" = list(
    "S" = c(1,2,3),
    "I" = c(0,0,1)
  )
)
age_groups <- c("0-5","5-19","19+")
create_age_initial_conditions(sim_scenario,age_groups)
}
