% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_model.R
\name{get_reactions_and_rates}
\alias{get_reactions_and_rates}
\title{convert reactions list to transitions and rates}
\usage{
get_reactions_and_rates(reactions)
}
\arguments{
\item{reactions}{list}
}
\value{
list of rates and transitions
}
\description{
convert reactions list to transitions and rates
}
