% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_model.R
\name{new_stochastic_model}
\alias{new_stochastic_model}
\title{constructor for \code{stochastic_model} class}
\usage{
new_stochastic_model(reactions, sim_scenario)
}
\description{
constructor for \code{stochastic_model} class
}
