% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic_model.R
\name{plot.abc_stochastic_model}
\alias{plot.abc_stochastic_model}
\title{type of plot either "prior_posterior" or "simulations"}
\usage{
\method{plot}{abc_stochastic_model}(x, ...)
}
\arguments{
\item{x}{An object of class \code{abc_stochastic_model}.}

\item{...}{list of additional plot arguments.
\code{param} - if \code{type} is "prior_posterior" then parameter to plot
\code{state} - if \code{type} is "simulations" then state to plot}
}
\description{
type of plot either "prior_posterior" or "simulations"
}
