% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_contact_matrix}
\alias{plot_contact_matrix}
\title{Plot a Contact Matrix}
\usage{
plot_contact_matrix(matrix_data, levels = NULL)
}
\arguments{
\item{matrix_data}{A square matrix where rows and columns represent different groups, and values indicate contact rates.}

\item{levels}{list of labels to order matrix}
}
\value{
A ggplot2 object representing the heatmap of the contact matrix.
}
\description{
This function visualizes a contact matrix using a heatmap.
}
\examples{
contact_matrix <- matrix(runif(25, 0, 1), nrow = 5, dimnames = list(LETTERS[1:5], LETTERS[1:5]))
plot_contact_matrix(contact_matrix)

}
