% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_stochastic_model.R
\name{scenario_stochastic_model}
\alias{scenario_stochastic_model}
\title{Create a set of scenarios that can
be incorporated into \code{\link[=projection_stochastic_model]{projection_stochastic_model()}}}
\usage{
scenario_stochastic_model(sm, parameters = NULL, states = NULL)
}
\arguments{
\item{sm}{Object of class \code{stochastic_model}.}

\item{parameters}{A data frame or tibble of prior parameter values. Each row
should correspond to one simulation, and columns should match the parameter
names expected by \code{update_parameters()}. If \code{NULL} then no parameters will
be updated an are all inherited from \code{sm}}

\item{states}{A data frame where each row represents an initial state. If
\code{NULL} then this is generated from the initial states in \code{sm}}
}
\description{
Create a set of scenarios that can
be incorporated into \code{\link[=projection_stochastic_model]{projection_stochastic_model()}}
}
\seealso{
\code{\link[=abc_stochastic_model]{abc_stochastic_model()}}
}
