% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_stochastic_model.R, R/stochastic_model.R
\name{update_parameters}
\alias{update_parameters}
\alias{update_state}
\alias{stochastic_model}
\alias{run_sim}
\title{update parameters for a stochastic_model}
\usage{
update_parameters(x, ..., params = list())

update_state(x, state)

stochastic_model(reactions, sim_scenario)

run_sim(x, tf = NULL)
}
\arguments{
\item{x}{An object of class \code{stochastic_model}.}

\item{...}{parameters to update}

\item{params}{parameters to update passed in as a list}

\item{state}{data.frame}

\item{reactions}{list of state transitions. Each reaction is itself a list
that requires a \code{transition} vector and a \code{rate} vector}

\item{sim_scenario}{list that defines the scenario including \code{params},
\code{initial_states} and \code{sim_args} that must at least include simulation time
\code{T}}

\item{tf}{maximum time}
}
\value{
An object of class \code{stochastic_model}.
}
\description{
update parameters for a stochastic_model

update initial state for a stochastic_model

Create a stochastic model

run stochastic model using adaptive tau set algorithm
}
\examples{
\dontrun{
sm <- update_parameters(sm,beta=0.2)
}
reactions <- list(
  infection = list(
    transition = c("I" = +1),
    rate = function(x,p,t){p$beta})
)
example_scenario <- list(
  params = list(beta = 0.1),
  initial_states = list(I = 0),
  sim_args = list(T = 10)
)
stochastic_model(reactions,example_scenario)

}
