% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_and_reactions.R
\name{vaccinate_initial_conditions}
\alias{vaccinate_initial_conditions}
\title{change vaccination coverage in initial conditions}
\usage{
vaccinate_initial_conditions(sim_scenario, vaccine_rate)
}
\arguments{
\item{sim_scenario}{output of \code{load_model_params}}

\item{vaccine_rate}{numeric proportion vaccinated. Can be a numeric
vector if using age groups. If named then vaccinates according to each
age group with format "S_\{age_group\}"}
}
\description{
change vaccination coverage in initial conditions
}
