% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_and_reactions.R
\name{load_model_params}
\alias{load_model_params}
\title{load model parameters}
\usage{
load_model_params(file = NULL)
}
\arguments{
\item{file}{file path to yaml}
}
\description{
load model parameters
}
